@@ -18,54 +18,62 @@
     "start": "vite --host"
   },
   "devDependencies": {
-    "@cypress/code-coverage": "3.9.12",
+    "@babel/core": "^7.18.5",
+    "@babel/preset-env": "^7.18.2",
+    "@cypress/code-coverage": "3.10.0",
+    "@jest/globals": "^28.1.1",
+    "@jest/types": "^28.1.1",
     "@testing-library/dom": "8.13.0",
     "@testing-library/jest-dom": "5.16.4",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/react-hooks": "7.0.2",
-    "@testing-library/user-event": "13.5.0",
-    "@types/jest": "27.4.1",
-    "@types/node": "17.0.18",
-    "@types/react": "17.0.45",
-    "@types/react-dom": "17.0.17",
-    "@typescript-eslint/eslint-plugin": "5.17.0",
-    "@typescript-eslint/parser": "5.17.0",
+    "@testing-library/react": "13.3.0",
+    "@testing-library/user-event": "14.2.0",
+    "@types/jest": "^28.1.1",
+    "@types/node": "17.0.43",
+    "@types/react": "18.0.12",
+    "@types/react-dom": "18.0.5",
+    "@types/testing-library__jest-dom": "^5.14.3",
+    "@typescript-eslint/eslint-plugin": "5.28.0",
+    "@typescript-eslint/parser": "5.28.0",
     "@vitejs/plugin-legacy": "1.8.2",
-    "@vitejs/plugin-react-refresh": "1.3.6",
+    "@vitejs/plugin-react": "^1.3.2",
     "autoprefixer": "10.4.7",
-    "eslint": "8.12.0",
+    "babel-loader": "^8.2.5",
+    "eslint": "8.17.0",
     "eslint-import-resolver-typescript": "2.7.1",
     "eslint-plugin-cypress": "2.12.1",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-jest": "26.1.5",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jest": "26.5.3",
     "eslint-plugin-jest-dom": "4.0.2",
     "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
+    "eslint-plugin-react": "7.30.0",
+    "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-simple-import-sort": "7.0.0",
-    "eslint-plugin-testing-library": "5.2.1",
-    "husky": "7.0.4",
+    "eslint-plugin-testing-library": "5.5.1",
+    "husky": "8.0.1",
     "identity-obj-proxy": "3.0.0",
     "istanbul-lib-coverage": "3.2.0",
-    "jest": "27.5.1",
+    "jest": "28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
     "jest-html-reporters": "3.0.9",
     "nyc": "15.1.0",
     "p-limit": "4.0.0",
     "postcss": "8.4.14",
     "postcss-import": "14.1.0",
     "postcss-nesting": "10.1.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "sass": "1.49.11",
-    "ts-jest": "27.1.5",
-    "typescript": "4.6.4",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "sass": "1.52.3",
+    "ts-jest": "28.0.5",
+    "ts-node": "^10.8.1",
+    "typescript": "4.7.3",
     "vite": "2.9.12",
-    "vite-plugin-istanbul": "2.5.1",
-    "vite-plugin-pwa": "0.11.13",
-    "web-vitals": "2.1.4"
+    "vite-plugin-istanbul": "2.8.0",
+    "vite-plugin-pwa": "0.12.0",
+    "web-vitals": "2.1.4",
+    "webpack": "^5.73.0"
   },
   "optionalDependencies": {
-    "cypress": "^9.5.0",
+    "cypress": "^10.1.0",
     "start-server-and-test": "^1.14.0"
   }
 }
