@@ -1,13 +1,14 @@
 import '@/main.scss';
 
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { App } from '@/App';
 import { reportWebVitals } from '@/reportWebVitals';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
-);
+const container = document.getElementById('root');
+
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!);
+root.render(<App />);
 
 reportWebVitals();
